+++ src/components/trailsList.js
@@ -2,14 +2,54 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import TrailItem from './trailItem.js'
+import Dialog from './dialog.js'
 
 import '../styles/trailList.css'
 
-const TrailsList = ({ trails }) => {
+const TrailsList = ({
+  trails,
+  error,
+  onStartTrail,
+  onStartTrailAccept,
+  onStartTrailCancel
+}) => {
+  let errorDialog = ''
+  if (error !== undefined) {
+    const dialogToolbar = [
+      (
+        <button
+          key='c'
+          onClick={onStartTrailCancel}
+        >
+          Abbrechen
+        </button>),
+      (
+        <button
+          key='o'
+          onClick={() => onStartTrailAccept(error.selectedTrail.uuid)}
+        >
+        Trail {error.selectedTrail.name} Starten
+        </button>)
+    ]
+    errorDialog = (
+      <Dialog title='Achtung' open={error !== undefined} tools={dialogToolbar}>
+        <p>Der Trail {error.currentTrail.name} läuft bereits</p>
+        <p>Willst du wirklich einen neuen Trail starten</p>
+      </Dialog>
+    )
+  }
   return (
     <section className='page--trails_list'>
+      {errorDialog}
       <ul className='trails_list'>
-        {trails.map((trail) => <TrailItem key={trail.uuid} {...trail} />)}
+        {trails.map(
+          (trail) =>
+            <TrailItem
+              key={trail.uuid}
+              onStartTrail={onStartTrail}
+              {...trail}
+            />
+        )}
       </ul>
     </section>
   )
@@ -22,7 +62,11 @@ TrailsList.propTypes = {
       description: PropTypes.string,
       uuid: PropTypes.string
     })
-  )
+  ),
+  error: PropTypes.object,
+  onStartTrail: PropTypes.func,
+  onStartTrailAccept: PropTypes.func,
+  onStartTrailCancel: PropTypes.func
 }
 
 export default TrailsList