+++ src/epics/trails.epics.js
@@ -1,8 +1,16 @@
-import { LOAD_TRAILS_START, LOAD_TRAILS_SUCCESS, LOAD_TRAILS_FAIL } from '../constants/trails.constants.js'
-import { mergeMap, catchError } from 'rxjs/operators'
+import {
+  LOAD_TRAILS_START,
+  LOAD_TRAILS_SUCCESS,
+  LOAD_TRAILS_FAIL,
+  START_TRAIL_BEGIN,
+  START_TRAIL_SUCCESS,
+  START_TRAIL_REJECT
+} from '../constants/trails.constants.js'
+import { mergeMap, catchError, map, mapTo } from 'rxjs/operators'
 import { of } from 'rxjs'
-import { ofType } from 'redux-observable'
+import { ofType, combineEpics } from 'redux-observable'
 import { normalizeTrails } from '../utils/transforms/trails.transforms.js'
+import { push } from 'connected-react-router'
 
 export const loadTrailsEpic = (action$, state$, { fetchJSON }) => action$.pipe(
   ofType(LOAD_TRAILS_START),
@@ -15,9 +23,49 @@ export const loadTrailsEpic = (action$, state$, { fetchJSON }) => action$.pipe(
     }
   }),
   catchError((e) => {
+    return of({
+      type: LOAD_TRAILS_FAIL
+    })
+  })
+)
+
+export const startTrailEpic = (action$, state$, dep) => action$.pipe(
+  ofType(START_TRAIL_BEGIN),
+  map(action => {
+    if (state$.value.trails.current_trail !== undefined) {
+      return {
+        type: START_TRAIL_REJECT,
+        trailId: action.trailId
+      }
+    }
+    return {
+      type: START_TRAIL_SUCCESS,
+      trailId: action.trailId
+    }
+  })
+  /* ,
+   catchError((e) => {
     console.error(e)
     return of({
       type: LOAD_TRAILS_FAIL,
     })
-  })
+  }) */
+)
+
+export const startTrailSuccessEpic = (action$, state$, dep) => action$.pipe(
+  ofType(START_TRAIL_SUCCESS),
+  mapTo(push('/mytrail'))
+  /* ,
+   catchError((e) => {
+    console.error(e)
+    return of({
+      type: LOAD_TRAILS_FAIL,
+    })
+  }) */
+)
+
+export default combineEpics(
+  startTrailEpic,
+  loadTrailsEpic,
+  startTrailSuccessEpic
 )